Windows: fix "corrupt installation" at new year
Bazel on Windows is now consistent with Bazel on
Unixes, by setting the mtimes of embedded binaries
to 10 years in the future.

Before this change, on Windows, Bazel used to set
these mtimes to CURRENT_YEAR + 10, January 1st.

This meant that if a user ran Bazel on 2017/12/29,
then on Unix Bazel set the mtimes to 2027/12/29
but on Windows it set them to 2027/01/01.

If the user then ran Bazel in the same workspace
on 2018/01/02, on Unixes it worked fine, but on
Windows it detected that the embedded binaries'
mtime is older than 2018/01/01, and reported a
"corrupt installation" error.

Fixes https://github.com/bazelbuild/bazel/issues/4378

Change-Id: I3457bdc360a62a279d1d08c9a69997929f2067dd

Closes #5385.

Change-Id: I3457bdc360a62a279d1d08c9a69997929f2067dd
PiperOrigin-RevId: 200395493