Save nested set size after initial expand to avoid GC from resizing list when replaying.
We stash the size in the same field as the order using bit packing. The small additional cost of masking out the int should be less than the resizing when replaying large nested sets.

The upper size bound on nested sets will be 512M objects as a result. If Java had unsigned ints we could push it to 1G.

RELNOTES: None
PiperOrigin-RevId: 200417105