Don't rerun the command if it failed.
While the original intent was to provide better messaging, the truth is things
like ibtool and actool can error because of the state of CoreServices when
targeting anything other than macOS. It appears this attempt to validate
the sdk argument ends up succeeding in some cases, meaning it doesn't
fail quickly, it does a fair mount of work, which we promptly throw away
and still return the original error result.

Rather then trying to expand this into some even more completely retry logic
it seems better to back away from this attempted messaging instead.

Looking at at history the better messaging stems from why it was recommended
that folks also pass the SDK version along with the Xcode version, but we've
gone back to just saying specific the Xcode version, so need for this
improved messaging has also been reduced.

PiperOrigin-RevId: 200422952