Bazel binary size: add a regression test
Add a regression test for the size of the Bazel
binary, by asserting the number of embedded tools.

Sudden, unexpectedly large changes in the number
of embedded tools can be indicative of an
unintentional addition/removal of embedded tools
and unexpected growth/shrinkage of the Bazel
binary.

Fixes https://github.com/bazelbuild/bazel/issues/5378

Change-Id: I7880f4544c560eb627ef5fb8a55ff1b377ec156b