Remove the removeHandler/releaseHandler pairs
They weren't actually doing anything. With default flags, createHandler never
returns a FancyTerminalEventHandler, because that always gets wrapped in a
RateLimitingEventHandler. Even so, under normal circumstances, the fancy
terminal event handler always leaves the terminal in a valid state at the end
of each printed line. We can only end up in a bad state if the blaze server is
killed, but even then, this code wasn't doing anything because it didn't get to
run.

The only way to guarantee that the terminal is always in a valid state is to
make sure that every stdout/stderr buffer we send to the client ends with a
reset code sequence. The experimental ui handler is better at sending only
complete buffers (we were previously using an auto-line flushing output stream,
which didn't allow the event handler to reason about how much stuff got
actually printed in a single write call).

Anyway, I think we're in a better place now, and this code wasn't run, and if
we want to fix it properly (if the switch to experimental_ui isn't sufficient),
then we need to fix it elsewhere.

PiperOrigin-RevId: 200699609