Don't propagate empty asset containers
If asset containers are empty, simply discard them rather than propagating them up the dependency graph. We don't currently use merging output except at the top level, so we don't need any of this information. (If/when asset merging is redone, we'd need to redo this code anyway.)

RELNOTES: none
PiperOrigin-RevId: 200721055