Add a utility function to accept windows-style envvars in paths.
For paths passed to Bazel on the command line, the shell expands these variables, but for hardcoded defaults, we must make the library call ourselves.

We do not add similar support in Posix systems, where it is less common to rely on standard path-related variables.

Prerequisit for issue #4502.

RELNOTES: None.
PiperOrigin-RevId: 201183214