Add a minimum_os_version flag that is platform agnostic.
This corresponds to the os version in the clang target triple, though clang
does have corresponding os_version_min flags for Apple platforms.

Currently, Bazel has --XX_minimum_os flags for XX values corresponding to Apple
platforms.

This flag is the generic version of that (and can and will be set after the
Apple split transition, and will eventually be set in the Android split as
well) to support various Android API level improvements for native code.

PiperOrigin-RevId: 201453961