Windows,Bazel client: check embedded tools faster
The Bazel client on Windows is now 50% faster to
check the embedded tools than it was before.

Checking the extracted binaries now takes 133ms
on average instead of the previous 294ms.

Previously the client performed the same tasks
multiple times while trying to determine if a path
was a good extracted binary. (E.g. converted the
path to Windows format multiple times, checked if
it was a directory twice, opened the path twice.)

Now the client performes these tasks only once.