Convert directDeps to a map of SkyValues
SkyFunctionEnvironment only cares about directDeps' values, not other
NodeEntry data.

This reduces the space of code which could be sensitive to nodes which
transition from done to dirty during evaluation.

To prevent check-then-act races in the refactored code (and only there;
other code will be fixed in future refactorings), instead of checking
deps' isDone() methods before accessing their value, allow
getValueMaybeWithMetadata to be called when not done, and have it return
null when not done.

(Note that done->dirty node transitions during evaluation are planned,
but not yet possible.)

RELNOTES: None.
PiperOrigin-RevId: 202518781