Store SkyValues for new deps from the graph intra-environment
This ensures that if a SkyFunction read a value (or its absence) for a
dep during its evaluation, subsequent requests for that dep's value (or
its absence) provide the same result.

This does not yet necessarily apply to the process of collecting events
and posts from a node's deps, which will be considered in a future
refactoring.

This CL adds SkyFunctionEnvironment#removeUndoneNewlyRequestedDeps
because the prior strategy for removing undone deps for done parents, by
re-requesting the dep's node from the graph and checking its doneness,
could lead to deps being dropped from parents if those deps transitioned
from done to dirty as the parent completes.

Minor cleanup to the bubbleErrorInfo field, which is nullable, and now
documented.

(Note that done->dirty node transitions during evaluation are planned,
but not yet possible.)

RELNOTES: None.
PiperOrigin-RevId: 202577098