Reuse previously stored SkyValues during event/post collection
If a SkyFunction read a value (or its absence) from the graph during its
evaluation, that value will be used to compute the event and post
metadata for that evaluation.

This CL modifies the assertion strategy for this code. Previously,
registered deps which were not done would have gone undetected, and
their events/posts skipped.

This CL also makes a few minor changes that make SkyFunctionEnvironment
more consistent:
 - Deps not already in previouslyRequestedDepsValues are added to
 newlyRequestedDeps regardless of whether evaluation was in
 error bubbling or whether the dep was done.
 - Previously requested deps of an inflight node are prefetched (by
 passing them to SkyFunctionEnvironment's ctor) during error bubbling in
 the same way as they are during normal eval or cycle checking.
 - Minor signature and documentation adjustments.

RELNOTES: None.
PiperOrigin-RevId: 202672709