bep: introduce BuildEventArtifactUploaderFactory
There can be multiple BuildEventTransports active at
the same time and we need to ensure that each transport
gets its own BuildEventArtifactUploader as these transports
might have different lifecycles.

We do that by introducing another level of indirection via
the BuildEventArtifactUploaderFactory. BlazeModules now
register a factory object instead of an uploader.

In addition, the BuildEventArtifactUploader gets a shutdown()
method that allows to free any resources associated with it.

PiperOrigin-RevId: 203752092