Remove the "slow read" warning when using multi-threaded digesting.
Calculating the throughput of a digest operation and using it to assess whether
I/O is slow or not only makes sense when we're doing sequential I/O. With
parallel hashing, it's expected that individual operations are slow due to how
the scheduler works, but it doesn't mean that the overall progress is slow.

RELNOTES: None.
PiperOrigin-RevId: 204115311