C++: Remove CcDynamicLibrariesForRuntime.
This was providing runtime libraries. This provider is redundant because the
same libraries are in CcLinkParamsStore.

Note to sheriff: Shouldn't break anything. If any, breakages would most likely be in Go test rules. I will look at the nightly tomorrow anyway. The breakages in the linked TGP re-run are not caused by this change, later re-runs have unrelated changes.

TESTED=[]
RELNOTES:none
PiperOrigin-RevId: 204738136