bes: fix bug where the bes upload would lose events
If a local file upload fails, then the build event
gets lost. This was the case because of a bug in the
implementation where local file upload errors are
retried by the BES retrier. This is the root cause
for the "Non-consecutive sequence number" errors we
have been seeing recently.

RELNOTES: None
PiperOrigin-RevId: 204906550