C++: Removes check for host configuration
The C++ Skylark API is protected by a flag and whitelist. The flag's value
should be true by default for non-Bazel. This is set with an invocation policy. In
Bazel the default value is false. When this was checked in I couldn't find a
way to use the default value in the host configuration, so when checking the
flag we ignored the value for these cases.

In this CL we copy the target config value to the host config.

RELNOTES:none
PiperOrigin-RevId: 205059612