Load toolchains with the target configuration
A toolchain might specify core libraries to be compiled
in the target configuration (e.g. libc or for my precise
case Rust' libprotobuf). With the current code, the compilation
would fail despite the toolchain specifying 'target' as the
transition configuration for those dependencies.

This was detected when trying to rebase the protobuf PR for
rules_rust and examples where failing because protobufs where
compiled with the host libprotobuf but the target example was
compiled with the target libprotobuf. Rustc detected a version.

/review @katre