Implement a way to do include validation as a part of input discovery. This makes it possible to disable .d file scanning when input discovery is used without allowing the usage of undeclared headers.
The way this is implemented relies on having a sand-boxed or remote execution
environment and simply removes undeclared files from discovered inputs. As a
result, the compiler cannot see them and can diagnose missing headers.

The input discovery itself cannot (usually) diagnose undeclared headers as it
is often implemented to be an over-approximation. It needs to find all used
headers, but it is allowed to find more. Diagnosing these additional headers
would not be useful.

RELNOTES: None.
PiperOrigin-RevId: 205628312