Use an interface to convert from nanos to millis since epoch
The conversion approach we were previously using is not stable - the resulting
offsets can differ based on what other things are going on on the same machine
at the same time.

By using an interface and passing it to the relevant places (and only computing
the offset once), we ensure that all conversions are consistent with each other.

PiperOrigin-RevId: 205787309