remote: fix race on download error. Fixes #5047
For downloading output files / directories we trigger all
downloads concurrently and asynchronously in the background
and after that wait for all downloads to finish. However, if
a download failed we did not wait for the remaining downloads
to finish but immediately started deleting partial downloads
and continued with local execution of the action.

That leads to two interesting bugs:
 1) The cleanup procedure races with the downloads that are still
    in progress. As it tries to delete files and directories, new
    files and directories are created and that will often
    lead to "Directory not empty" errors as seen in #5047.
 2) The clean up procedure does not detect the race, succeeds and
    subsequent local execution fails because not all files have
    been deleted.